bridge/mqtt/requirements.txt
paho-mqtt
azure-iot-device
